moving out some go files
