+++ client/cli/exercise.go
