+++ client/cli/formatter.go
