+++ client/cli/formatter_test.go
