+++ client/cli/user.go
